requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
langchain==0.0.335
openai==1.3.5
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
fastapi-websocket-rpc==0.1.16
reportlab==4.0.7
python-dotenv==1.0.0